Confirm before deleting Answers and Questions

+++ crowbot/backend/management/commands/delete_questions.py
@@ -6,8 +6,13 @@ class Command(BaseCommand):
     help = 'Empty the Question and Answer tables'
 
     def handle(self, *args, **options):
-        self.stdout.write('Deleting questions and answers')
-        for q in Question.objects.all():
-            q.delete()
-        for a in Answer.objects.all():
-            a.delete()
+        self.stdout.write('Delete all questions and answers? (y/n)')
+        ans = input()
+        if ans and ans[0] == 'y':
+            self.stdout.write('Deleting questions and answers')
+            for q in Question.objects.all():
+                q.delete()
+            for a in Answer.objects.all():
+                a.delete()
+        else:
+            self.stdout.write('Aborting')
